Remove code that's not needed for this plugin
