--- openedx_blockstore_relay/models.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Database models for openedx_blockstore_relay.
-"""
-
-from __future__ import absolute_import, unicode_literals
