--- openedx_blockstore_relay/settings/private.py.sample
@@ -1,5 +1,0 @@
-"""
-Copy this file to openedx_blockstore_relay/settings/private.py
-
-and override any local settings here.
-"""
