--- openedx_blockstore_relay/settings/production.py
@@ -1,9 +1,0 @@
-"""
-Production settings for the openedx-blockstore-relay app.
-"""
-
-from __future__ import absolute_import, unicode_literals
-
-from .common import *  # pylint: disable=wildcard-import
-
-# TODO: grab settings from ENV_TOKENS so they can be configured from ansible
