--- openedx_blockstore_relay/urls.py
@@ -1,12 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-URLs for openedx_blockstore_relay.
-"""
-from __future__ import absolute_import, unicode_literals
-
-from django.conf.urls import url
-from django.views.generic import TemplateView
-
-urlpatterns = [
-    url(r'', TemplateView.as_view(template_name="openedx_blockstore_relay/base.html")),
-]
