--- test_settings.py
@@ -1,45 +1,0 @@
-"""
-These settings are here to use during tests, because django requires them.
-
-In a real-world use case, apps in this project are installed into other
-Django applications, so these settings will not be used.
-"""
-
-from __future__ import absolute_import, unicode_literals
-
-from os.path import abspath, dirname, join
-
-
-def root(*args):
-    """
-    Get the absolute path of the given path relative to the project root.
-    """
-    return join(abspath(dirname(__file__)), *args)
-
-
-DATABASES = {
-    'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': 'default.db',
-        'USER': '',
-        'PASSWORD': '',
-        'HOST': '',
-        'PORT': '',
-    }
-}
-
-INSTALLED_APPS = (
-    'django.contrib.auth',
-    'django.contrib.contenttypes',
-    'openedx_blockstore_relay',
-)
-
-LOCALE_PATHS = [
-    root('openedx_blockstore_relay', 'conf', 'locale'),
-]
-
-ROOT_URLCONF = 'openedx_blockstore_relay.urls'
-
-SECRET_KEY = 'insecure-secret-key'
-
-BLOCKSTORE_API_URL = 'http://localhost:8888/api/v1'
